added vehicle exclusion based on objective weights

+++ EADARP/Algorithms.h
@@ -100,11 +100,11 @@ namespace algorithms
 
 		Solution exchangeConsecutive(Solution s, NeighborChoice strategy);
 
-		Solution WorstRemoval(Solution s, double removal_ratio, float randomness);
+		Solution WorstRemoval(Solution s, double removal_ratio, double randomness);
 
 		Solution GreedyInsertion(Solution s);
 
-		Solution Destroy(Solution s, double removal_ratio, float randomness, Instance::Objective objective);
+		Solution Destroy(Solution s, double removal_ratio, double randomness, Instance::Objective objective);
 		
 		Solution Repair(Solution s,Instance::Objective objective);
